requirements.txt
asgiref==3.2.10
coverage==5.2
Django==3.0.8
djangorestframework==3.11.0
Markdown==3.2.2
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
